% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{evaluate_imputation}
\alias{evaluate_imputation}
\title{print_it_stats <- function(oob, n_oob, s_oob, dif, n_diff, sn_diff, trim, fc) {
}}
\usage{
evaluate_imputation(old, current)
}
\description{
print_it_stats <- function(oob, n_oob, s_oob, dif, n_diff, sn_diff, trim, fc) {
}
}
