% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Single imputation}
\usage{
imputation(
  data,
  workers = 1,
  mtry = floor(ncol(data)/3),
  init = init_mean_freq,
  stop_crit = c("delta", "oob"),
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} or a \code{matrix} to perform the imputation on,
missing values has to be \code{NA}.}

\item{workers}{Number of parallel workers to use when building the forest}

\item{mtry}{Number of columns to use in each tree}

\item{init}{A function for initialization of missing values in the data}

\item{...}{Additional arguments to \link[ranger:ranger]{ranger::ranger}.}
}
\value{
a \code{data.frame} with \code{NA} values replaced by imputed values.
}
\description{
Performs a single imputation run and returns the data with NA values replaced
by imputed values.
}
\examples{
iris_miss <- iris
iris_miss[
  matrix(
    c(
        sample.int(nrow(iris_miss), 50),
        sample.int(ncol(iris_miss), 50, replace = TRUE)
      ), ncol = 2)
  ] <- NA
imputation(iris_miss)
}
