% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mean_freq.R
\name{init_mean_freq}
\alias{init_mean_freq}
\title{Initialize The Data With Missing Values Using The Mean Or Frequency Imputation}
\usage{
init_mean_freq(data)
}
\arguments{
\item{data}{A data matrix to replace missing values in}
}
\value{
The input data with missing values replaced with the mean (\code{numerical})
or the label with the highest occurrence (\code{categorical}).
}
\description{
Replace data matrix with the mean for \code{numerical} data and the most frequent
instance for categorical data (\code{character} or \code{factor})
}
\examples{
df <- data.frame(
  a = c(runif(10), rep(NA, 5)),
  b = c(letters[sample.int(26, 10)], rep(NA, 5))
)
init_mean_freq(df)
}
