% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_freq_impute.R
\name{mean_freq_impute}
\alias{mean_freq_impute}
\title{Mean or Frequency Imputation}
\usage{
mean_freq_impute(x)
}
\arguments{
\item{x}{The vector to replace missing values in}
}
\value{
A vector with missing values replaced by the mean or most frequent observation
}
\description{
Replaces \code{NA} values with the mean for \code{numeric} or the most frequent observation for \code{character}.
For \code{character}s, ties are broken uniformly at random
}
\examples{
mean_freq_impute(c(runif(10), NA))
}
